move dirs and modinfo to common
