+++ tools/common/dirs/dirs_test.go
