+++ tools/common/modinfo/modinfo.go
@@ -12,9 +12,9 @@ import (
 	"strings"
 
 	"github.com/Azure/azure-sdk-for-go/tools/common/delta"
+	"github.com/Azure/azure-sdk-for-go/tools/common/dirs"
 	"github.com/Azure/azure-sdk-for-go/tools/common/exports"
 	"github.com/Azure/azure-sdk-for-go/tools/common/report"
-	"github.com/Azure/azure-sdk-for-go/tools/internal/dirs"
 	"github.com/Masterminds/semver"
 )
