+++ tools/common/modinfo/modinfo_test.go
