+++ tools/profileBuilder/cmd/list.go
@@ -14,8 +14,8 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/Azure/azure-sdk-for-go/tools/internal/dirs"
-	"github.com/Azure/azure-sdk-for-go/tools/internal/modinfo"
+	"github.com/Azure/azure-sdk-for-go/tools/common/dirs"
+	"github.com/Azure/azure-sdk-for-go/tools/common/modinfo"
 	"github.com/Azure/azure-sdk-for-go/tools/profileBuilder/model"
 	"github.com/spf13/cobra"
 )
